+++ torchao/quantization/qat/__init__.py
@@ -2,12 +2,7 @@ from .api import (
     ComposableQATQuantizer,
 )
 from .linear import (
-    disable_4w_fake_quant,
-    disable_8da4w_fake_quant,
-    enable_4w_fake_quant,
-    enable_8da4w_fake_quant,
     Int4WeightOnlyQATQuantizer,
-    Int8DynActInt4WeightQATLinear,
     Int8DynActInt4WeightQATQuantizer,
 )
 from .embedding import (
@@ -15,13 +10,8 @@ from .embedding import (
 )
 
 __all__ = [
-    "disable_4w_fake_quant",
-    "disable_8da4w_fake_quant",
-    "enable_4w_fake_quant",
-    "enable_8da4w_fake_quant",
     "ComposableQATQuantizer",
     "Int4WeightOnlyQATQuantizer",
     "Int4WeightOnlyEmbeddingQATQuantizer"
     "Int8DynActInt4WeightQATQuantizer",
-    "Int8DynActInt4WeightQATLinear",
 ]
