+++ torchao/quantization/qat/images/qat_diagram.png
