+++ torchao/quantization/qat/linear.py
